+++ TRAP/src/ImageLoader/PortableNetworkGraphics/PNGImage.h
@@ -9,15 +9,42 @@ namespace TRAP::INTERNAL
 	class PNGImage final : public Image
 	{
 	public:
+		/// <summary>
+		/// Constructor.
+		/// </summary>
+		/// <param name="filepath">File path of the Image to load.</param>
 		explicit PNGImage(std::string filepath);
+		/// <summary>
+		/// Default Copy Constructor.
+		/// </summary>
 		PNGImage(const PNGImage&) = default;
+		/// <summary>
+		/// Default Copy Assignment Operator.
+		/// </summary>
 		PNGImage& operator=(const PNGImage&) = default;
+		/// <summary>
+		/// Default Move Constructor.
+		/// </summary>
 		PNGImage(PNGImage&&) = default;
+		/// <summary>
+		/// Default Move Assignment Operator.
+		/// </summary>
 		PNGImage& operator=(PNGImage&&) = default;
+		/// <summary>
+		/// Default Destructor.
+		/// </summary>
 		~PNGImage() = default;
 
+		/// <summary>
+		/// Retrieve the raw pixel data of the Image.
+		/// </summary>
+		/// <returns>Constant pointer to the raw pixel data.</returns>
 		const void* GetPixelData() const override;
-		uint32_t GetPixelDataSize() const override;
+		/// <summary>
+		/// Retrieve the size of the raw pixel data of the Image.
+		/// </summary>
+		/// <returns>Size of the raw pixel data in bytes.</returns>
+		uint64_t GetPixelDataSize() const override;
 
 	private:
 		struct RGBA
